created a dedicated component for work experiences, and added a tooltip for interactiveness
